CIV-6282 Refactor judge Decision handler test class
